+++ malcolm/modules/pmac/parts/brickpart.py
@@ -9,8 +9,11 @@ class BrickPart(builtin.parts.ChildPart):
     def setup(self, registrar):
         # type: (PartRegistrar) -> None
         super(BrickPart, self).setup(registrar)
+        # Hooks
         self.register_hooked(scanning.hooks.ReportStatusHook,
                              self.report_status)
+        self.register_hooked(scanning.hooks.PauseHook,
+                             self.resync)
 
     @add_call_types
     def report_status(self, context):
@@ -21,3 +24,20 @@ class BrickPart(builtin.parts.ChildPart):
         outputs_list = [(outputs >> i) & 1 for i in range(16)]
         controller_info = ControllerInfo(child.i10.value, outputs_list)
         return controller_info
+
+    @add_call_types
+    def reset(self, context):
+        # type: (builtin.hooks.AContext) -> None
+        super(BrickPart, self).reset(context)
+        self.resync(context)
+
+    @add_call_types
+    def resync(self, context):
+        # type: (scanning.hooks.AContext) -> None
+        # The GPIO is polled in the medium loop, so wait for it to update.
+        # Need this so that report status is accurate, but report status
+        # needs to be quick so can't do it there
+        # TODO: replace this with a poll now button
+        # child = context.block_view(self.mri)
+        # Medium loop is every 2s so 3s should be enough
+        context.sleep(3.0)